TheHarvester support and pwned check

+++ setup/setup.py
@@ -32,10 +32,17 @@ def shodanSetup():
 
 def downloadLibs():
     # Download theHarvester
-    print "Downloading theHarvester.py"
-    os.system('wget https://github.com/laramies/theHarvester/archive/master.zip -O ../lib/harvester.zip')
-    os.system('unzip ../lib/harvester.zip -d ../lib/')
-    os.system('mv ../lib/theHarvester-master ../lib/theharvester')
-    os.system('rm ../lib/harvester.zip')
+    if os.path.isfile('../lib/theharvester'):
+        print "Downloading theHarvester.py"
+        os.system('wget https://github.com/laramies/theHarvester/archive/master.zip -O ../lib/harvester.zip')
+        os.system('unzip ../lib/harvester.zip -d ../lib/')
+        os.system('mv ../lib/theHarvester-master ../lib/theharvester')
+        os.system('rm ../lib/harvester.zip')
+        os.system('mv ../lib/theharvester/lib/graphs.py ../lib/graphs.py')
+        os.system('mv ../lib/theharvester/lib/hostchecker.py ../lib/hostchecker.py')
+        os.system('mv ../lib/theharvester/lib/htmlExport.py ../lib/htmlExport.py')
+        os.system('mv ../lib/theharvester/lib/markup.py ../lib/markup.py')
+    else:
+        print "TheHarvester is installed."
 
 setup()
